Typescript convertion
